Removes the need for the markdownHTML-component, and deletes it
