+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import DocuPage from './components/docupage/docupage.js'
 import { Sidebar } from './components/sidebar/sidebar.js';
 import { DocCard } from './components/startpage/startpage.js';
 import { BrowserRouter as Router, Switch, Route} from "react-router-dom";
-import MarkdownHTML from './components/MarkdownHTML/MarkdownHTML.js'
 import './vippsstyle.css';
 import './index.css';
 import vipps_dev from "./img/vipps_dev.svg"
@@ -54,93 +54,6 @@ const Cards = () => (
     </div>
 )
 
-// The entire page is contained here
-class DocuPage extends React.Component {
-  constructor(props) {
-    super(props);
-    this.urls = {
-      ecom:
-        "https://raw.githubusercontent.com/vippsas/vipps-ecom-api/master/vipps-ecom-api.md",
-      login:
-        "https://raw.githubusercontent.com/vippsas/vipps-login-api/master/vipps-login-api.md",
-      invoice:
-        "https://raw.githubusercontent.com/vippsas/vipps-invoice-api/master/vipps-invoice-api.md"
-    };
-    this.state = {
-      fullText: "",
-      headers: []
-    };
-  }
-
-  componentDidMount() {
-    this.getContent().then(response =>
-      response.text().then(text => this.getHeaders(text))
-    );
-  }
-
-  // Fetches raw content from Github and puts it in the DocuPage state
-  getContent() {
-    return fetch(this.urls[this.props.doc]);
-  }
-
-  // Returns a HTML anchor from a given header
-  makeAnchor(string) {
-    return (
-      "#" +
-      string
-        .replace(new RegExp("[|&;:$%@<>()+,#]", "g"), " ")
-        .trim()
-        .replace(new RegExp(" ", "g"), "-")
-        .toLowerCase()
-      );
-    }
-
-    // Filters the content fetched from Github into headers
-    getHeaders(data) {
-        const originalMarkdown = (' ' + data).slice(1);
-        const lines = data.split("\n");
-        let navbarHeaders = []
-        let navbarHeader = {name: "", anchor: "", children: []}
-        lines.forEach((line) => {
-            if (line.startsWith("###")) {
-                return;
-            } else if (line.startsWith("##")) {
-                navbarHeader.children.push({name: line.replace("##", "").trim(),
-                    anchor: this.makeAnchor(line)});
-            } else if (line.startsWith("#")) {
-                navbarHeaders.push(navbarHeader);
-                navbarHeader = {name: "", anchor: "", children: []}
-                navbarHeader.name = line.replace("#", "").trim();
-                navbarHeader.anchor = this.makeAnchor(line);
-            } else {
-                return;
-            }
-        });
-        this.setState({
-            fullText: originalMarkdown,
-            // First element i navbarHeaders is an empty collection
-            // Second element is just the name of the documentation, not needed in navigation bar
-            // In case second header is not 'table of contents' do not exclude the second header
-            headers: navbarHeaders[2].name === "Table of contents" ? navbarHeaders.slice(3) : navbarHeaders.slice(2)
-        });
-    }
-
-    render() {
-        return (
-            <div className="container bold">
-              <div className="sidebar">
-                <Sidebar headers={this.state.headers} api={this.props.doc}/>
-              </div>
-              <div className="content">
-                <MarkdownHTML
-                  //url={"https://raw.githubusercontent.com/vippsas/vipps-ecom-api/master/vipps-ecom-api.md"}
-                  text={this.state.fullText}
-                />
-              </div>
-            </div>
-        )
-    }
-}
 
 ReactDOM.render(
     <StartPage/>,
